Split InnerContent from ContentRenderer
